[matchingImageCollection] Extract image pair IO to separate file
